requirements.txt
Flask==2.0.1
pycryptodome==3.10.1
python-dotenv==0.19.0 